feat: interfaces-b11-cli-work-006 - Verify contract/package gates and ownership bounds
Consolidate work-family parity tests to satisfy pkg-file-count, refresh CLI identity baselines after work metadata alignment, and regenerate staged contract artifacts so contracts-smoke passes.
